pyproject.toml
[project]
name = "invest"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "databento>=0.74.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "seaborn>=0.13.2",
]
